#RI-2332 - add virtualized list to monitor output
