+++ redisinsight/ui/src/components/monitor/Monitor/styles.module.scss
@@ -28,6 +28,15 @@
   }
 }
 
+.listWrapper {
+  @include euiScrollBar;
+
+  width: 100%;
+  height: 100%;
+  position: relative;
+  overflow: auto;
+}
+
 .content {
   @include euiScrollBar;
